requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
beautifulsoup4==4.9.3
bleach==3.3.0
brotlipy==0.7.0
bs4==0.0.1
certifi==2020.12.5
cffi @ file:///C:/ci/cffi_1625831756778/work
chardet==4.0.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.0.1
colorama==0.4.4
conda==4.10.3
conda-package-handling @ file:///C:/ci/conda-package-handling_1618262410900/work
cryptography @ file:///C:/ci/cryptography_1616769504165/work
cycler==0.10.0
docutils==0.16
Flask==2.0.1
future==0.18.2
gTTS==2.2.3
html5lib==1.1
idna==2.10
importlib-metadata==3.7.2
itsdangerous==2.0.1
Jinja2==3.0.1
keyring==23.0.0
kiwisolver==1.3.1
line-bot-sdk==1.19.0
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.2
menuinst==1.4.16
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184308/work
mkl-service==2.4.0
multidict==5.1.0
numpy @ file:///C:/ci/numpy_and_numpy_base_1634088717532/work
packaging==20.9
piexif==1.1.3
Pillow==8.4.0
playsound==1.2.2
prompt-toolkit==1.0.14
pycodestyle==2.7.0
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments==2.8.1
PyInquirer==1.0.3
pyngrok==5.0.5
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing==2.4.7
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python-dateutil==2.8.1
pywin32==228
pywin32-ctypes==0.2.0
PyYAML==5.4.1
readme-renderer==29.0
regex==2021.11.10
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.4.0
ruamel-yaml-conda @ file:///C:/ci/ruamel_yaml_1616016898638/work
six==1.15.0
soupsieve==2.2.1
tabulate==0.8.9
tinydb==4.5.2
toml==0.10.2
tqdm==4.59.0
twine==3.3.0
typing-extensions==3.10.0.0
urllib3==1.26.3
vboxapi==1.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
wincertstore==0.2
yarl==1.6.3
zipp==3.4.1
